+++ cdborm/errors.py
@@ -16,3 +16,5 @@ class CanNotOverwriteRelationVariable(Exception): pass
 class AlreadyAssigned(Exception): pass
 class BadValueType(Exception): pass
 class NoDbSelected(Exception): pass
+
+from CodernityDB.database import RecordNotFound, RecordDeleted