+++ History.md
@@ -9,6 +9,7 @@
   * get, all and save no raises NoDatabaseSelected when database == None.
   * default now can be used with 2 arguments or no arguments
   * fixed removing elements from databas (remove now delete element from cache)
+  * Added RecordNotFound and RecordDeleted to cdborm.errors
 
 0.1.0 / 2012-12-15
 ==================